Clock In Clock Out Functionality with Files
The Files are now working with the clock in and clock out functionality.
The program now remembers which employees have clocked in or out during
the program run. If the program is closed and opened again, the
employees are remembered now.

If an employee tries to clock in while already clocked in, the missed
punch screen is displayed. After entering data and submitting, they are
then clocked in.