Add Employee Now Effects Files
If the manager adds an employee, they are generated with an ID number of
the next number up. i.e. the file (if none is found) is created with 4
employees. If a fifth employee is added, their ID is 55555. A sixth
would be 66666. And so on...

After the program closes, again, these newly added employees are still
available and registered as being clocked in or clocked out.